+++ src/framework/theme/components/cdk/overlay/index.ts
@@ -5,4 +5,5 @@ export * from './overlay-container';
 export * from './overlay-trigger';
 export * from './mapping';
 export * from './position-helper';
-export * from './dynamic-overlay';
+export * from './dynamic/dynamic-overlay';
+export * from './dynamic/dynamic-overlay-handler';
